ndfc-fabric-exporter/requirements.txt
# Cisco NDFC API Client
requests>=2.31.0
urllib3>=2.0.0

# Ansible (compatible with ansible-core 2.18.8)
# Note: Only install ansible-core if you have it already installed
# The tool works with any ansible-core >= 2.15.0
# ansible-core>=2.15.0,<3.0.0

# YAML Processing (Python 3.13.2 compatible)
PyYAML>=6.0.1
ruamel.yaml>=0.18.0

# JSON Schema Validation (for nac-validate style validation)
jsonschema>=4.19.0

# Configuration Management
python-dotenv>=1.0.0

# CLI and Utilities (Python 3.13.2 compatible)
click>=8.1.7
rich>=13.7.0
tabulate>=0.9.0

# Logging
colorlog>=6.7.0

# Optional: for advanced validation
# pydantic>=2.0.0
